✨ 코드레빗 피드백 반영 : 소셜 로그인 딥링크 연동 및 내비게이션 구조 개선
- **내비게이션 구조 개편**: `LoginApp.kt`를 `navigation` 패키지로 이동하고, 중복 코드를 `NavEntryHelper.kt`로 분리하여 공통화
- **소셜 로그인 고도화**:
    - 딥링크를 통한 `socialToken` 및 `status` 전달 방식을 `navArgument`로 변경
    - `SocialAuthViewModel` 및 `SocialEntryScreen`에 딥링크 처리 로직 추가
    - `WelcomeSocialScreen`에 로그인 성공 콜백(`onLoginSuccess`) 연동 및 재시도 로직 구현
- **보안 및 설정 관리**: 하드코딩된 서버 URL을 `ServerConfig.kt` 상수로 분리
- **버그 수정 및 안정화**:
    - `rememberSaveable`을 사용하여 회원가입 요청 상태 유지 (리컴포지션 대응)
    - `social_auth_graph` 백스택 접근 시 예외 처리 추가
    - `UserRepositoryImpl`에서 발생한 `ApiError`를 상위 레이어로 전파하도록 수정
- **코드 정리**: `MainActivity` 및 `MainApp`에 흩어져 있던 소셜 로그인 처리 로직을 관련 모듈로 캡슐화 및 불필요한 주석 삭제
