+++ app/src/main/java/com/example/linku_android/MainApp.kt
@@ -71,16 +71,15 @@ import androidx.core.net.toUri
 import com.example.core.model.auth.LoginState
 import com.example.linku_android.curation.curationGraph
 import com.example.linku_android.deeplink.appLinkRoute
-import com.example.login.LoginApp
+import com.example.login.navigation.LoginApp
 
 
 
 @Composable
 fun MainApp(
     viewModel: MainViewModel,
-    socialLoginData: SocialLoginData? = null
-) {
 
+) {
 
     // 앱 실행 시 실행하여 이전 계정 기록 삭제
     LaunchedEffect(Unit) {
@@ -91,30 +90,6 @@ fun MainApp(
 
     // 로그인에서 사용할 뷰모델
     val loginViewModel: LoginViewModel = hiltViewModel()
-
-    // 소셜 로그인 딥링크 처리
-    LaunchedEffect(socialLoginData) {
-        socialLoginData?.let { data ->
-            Log.d("SOCIAL_LOGIN", "=== MainApp에서 소셜 로그인 처리 ===")
-            Log.d("SOCIAL_LOGIN", "provider: ${data.provider}")
-            Log.d("SOCIAL_LOGIN", "token 길이: ${data.token.length}")
-            loginViewModel.handleSocialLoginToken(data.token, data.provider)
-        }
-    }
-
-    // 소셜 로그인 성공 시 홈으로 이동
-    val loginState by loginViewModel.loginState.collectAsStateWithLifecycle()
-
-    LaunchedEffect(loginState) {
-        Log.d("SOCIAL_LOGIN", "loginState 변경: $loginState")
-        if (loginState is LoginState.Success && socialLoginData != null) {
-            Log.d("MainApp", "소셜 로그인 성공! 홈으로 이동")
-            navigator.navigate(NavigationRoute.Home.route) {
-                popUpTo(0) { inclusive = true }
-                launchSingleTop = true
-            }
-        }
-    }
 
     // 홈 화면에서 사용할 뷰모델
     val homeViewModel: HomeViewModel = hiltViewModel()
@@ -212,7 +187,7 @@ fun MainApp(
             val deps = remember {
                 EntryPointAccessors.fromApplication(app, SplashDeps::class.java)
             }
-            val loginVM: LoginViewModel = hiltViewModel()
+
 
 
             NavHost(
@@ -267,7 +242,7 @@ fun MainApp(
                                 autoLoginTried = true
 
                                 // refresh 있음 → 자동로그인 시도
-                                loginVM.tryAutoLogin(
+                                loginViewModel.tryAutoLogin(
                                     onSuccess = {
                                         navigator.navigate(NavigationRoute.Home.route) {
                                             popUpTo(NavigationRoute.Splash.route) { inclusive = true }
